Switch data format from pickle to cbor
serde_cbor seems to be a better maintained rust
crate than serde_pickle. The file size is a bit
smaller, but more importantly it does not have
the issue that the pickle crate had, where
restarts written from Python could not be read
back in on the Rust end. The only downside,
which is very minor, is that we require the
cbor2 Python module as a dependency.