Switch from using toml for config to yaml
This serde stuff is not perfect. CBOR was
a better choice for binary outputs, but it
is not robust with serde's tagging options
(neither was serde_pickle, but I'd gotten
it to work). Without tags, CBOR works great
but the toml fails! It cannot read/write
nested dicts in an easily readable format.
So we're back to yaml now, with the nested
dict representation for enum's.