Change the Products struct to use AgnosticPrimitive
This would have been a bad idea for serializing
products to CBOR, since AgnosticPrimitive is a
struct, not a tuple struct, it would have put
the keys into each array element. If switching
to a Python loader this will not be a problem.
Later on, AgnosticPrimitive can always be made
a tuple struct if necessary.