Add an option to enable the adaptive time step
By default, the code has the same behavior now as
it did before the last several commits introduced
the adaptive time step. I restored the CFL number
to 0.4 for the two jet problems, and only enabled
the adaptive stepping for the halo-kilonova setup
since it has sub-relativistic velocities and will
probably benefit from this feature.