Use interior mutability (RefCell) to cache table
This approach could panic if the problem methods
are called on multiple threads. However it looks
like that never happens, because the initial data
is generated on the main thread, and then the
boundary conditions are *also* sampled on the main
thread (see scheme::add_remove_blocks).

Warning! This is not tested yet. When first running
it please make sure the table is actually loaded
just once, and that errors such as a missing or
corrupted table file are handled and reported
correctly.